feat: add error and loading stages and poll for status
