+++ src/components/messenger/user-profile/zero-pro-panel/index.tsx
@@ -6,8 +6,11 @@ import { Plans } from './stages/Plans';
 import { Details } from './stages/Details';
 import { Payment } from './stages/Payment';
 import { Success } from './stages/Success';
+import { Loading } from './stages/Loading';
+import { Error } from './stages/Error';
 import { BottomSheet } from './bottom-sheet';
 import styles from './styles.module.scss';
+import { usePollZeroProActiveStatus } from './usePollZeroProActiveStatus';
 
 export enum ZeroProStage {
   PaymentPlan = 'payment-plan',
@@ -27,47 +30,6 @@ export interface BillingDetails {
   city: string;
   postalCode: string;
   country: string;
-}
-
-// Polling hook for subscription status
-function usePollZeroProActiveStatus(
-  shouldPoll: boolean,
-  onActive: () => void,
-  onTimeout: () => void,
-  timeoutMs = 20000
-) {
-  React.useEffect(() => {
-    if (!shouldPoll) return;
-    let isMounted = true;
-    let timeoutId: NodeJS.Timeout;
-    let start = Date.now();
-    async function poll() {
-      while (isMounted) {
-        const res = await fetch('/subscription/status?type=ZERO');
-        const data = await res.json();
-        const status = data.subscription?.status ?? data.status;
-        if (status === 'active') {
-          onActive();
-          return;
-        }
-        if (Date.now() - start > timeoutMs) {
-          onTimeout();
-          return;
-        }
-        await new Promise((resolve) => setTimeout(resolve, 2000));
-      }
-    }
-    poll();
-    return () => {
-      isMounted = false;
-      clearTimeout(timeoutId);
-    };
-  }, [
-    shouldPoll,
-    onActive,
-    onTimeout,
-    timeoutMs,
-  ]);
 }
 
 export const ZeroProPanel: React.FC<{ onClose: () => void }> = ({ onClose }) => {
@@ -126,17 +88,8 @@ export const ZeroProPanel: React.FC<{ onClose: () => void }> = ({ onClose }) => 
           {activeSheet === ZeroProStage.Payment && (
             <Payment billingDetails={billingDetails} onNext={handlePaymentNext} onBack={openDetails} />
           )}
-          {activeSheet === ZeroProStage.Loading && (
-            <div className={styles.LoadingContainer}>
-              <div className={styles.Spinner} />
-              <div className={styles.LoadingText}>Activating your subscription...</div>
-            </div>
-          )}
-          {activeSheet === ZeroProStage.Error && (
-            <div className={styles.ErrorContainer}>
-              <div className={styles.ErrorText}>{pollError}</div>
-            </div>
-          )}
+          {activeSheet === ZeroProStage.Loading && <Loading />}
+          {activeSheet === ZeroProStage.Error && <Error error={pollError} />}
           {activeSheet === ZeroProStage.Success && <Success onClose={closeSheet} />}
         </BottomSheet>
       )}
