+++ src/components/messenger/user-profile/zero-pro-panel/styles.module.scss
@@ -1,3 +1,5 @@
+@use '~@zero-tech/zui/styles/theme' as theme;
+
 .ZeroProPanel {
   flex-grow: 1;
   display: flex;
